complete DocPage refactor
