+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/index.tsx
@@ -9,7 +9,7 @@ import React, {useState} from 'react';
 import Layout from '@theme/Layout';
 import BackToTopButton from '@theme/BackToTopButton';
 import type {Props} from '@theme/DocPage/Layout';
-import DocPageLayoutAside from '@theme/DocPage/Layout/Aside';
+import DocPageLayoutSidebar from '@theme/DocPage/Layout/Sidebar';
 import DocPageLayoutMain from '@theme/DocPage/Layout/Main';
 
 import styles from './index.module.css';
@@ -24,7 +24,7 @@ export default function DocPageLayout({children}: Props): JSX.Element {
       <BackToTopButton />
       <div className={styles.docPage}>
         {sidebar && (
-          <DocPageLayoutAside
+          <DocPageLayoutSidebar
             sidebar={sidebar.items}
             hiddenSidebarContainer={hiddenSidebarContainer}
             setHiddenSidebarContainer={setHiddenSidebarContainer}
